Move logging out of the model module and reduce amount of logging levels
