+++ android/src/main/java/org/jellyfin/apiclient/logging/AndroidLogger.kt
@@ -1,7 +1,6 @@
 package org.jellyfin.apiclient.logging
 
 import android.util.Log
-import org.jellyfin.apiclient.model.logging.ILogger
 
 /**
  * ILogger implementation for Android. Uses the [android.util.Log] class.
@@ -13,6 +12,10 @@ class AndroidLogger(
 		return String.format(raw, *parameters)
 	}
 
+	override fun debug(msg: String, vararg parameters: Any) {
+		Log.d(tag, getFormattedString(msg, *parameters))
+	}
+
 	override fun info(msg: String, vararg parameters: Any) {
 		Log.i(tag, getFormattedString(msg, *parameters))
 	}
@@ -21,23 +24,7 @@ class AndroidLogger(
 		Log.e(tag, getFormattedString(msg, *parameters))
 	}
 
-	override fun warn(msg: String, vararg parameters: Any) {
-		Log.w(tag, getFormattedString(msg, *parameters))
-	}
-
-	override fun debug(msg: String, vararg parameters: Any) {
-		Log.d(tag, getFormattedString(msg, *parameters))
-	}
-
-	override fun fatal(msg: String, vararg parameters: Any) {
-		Log.wtf(tag, getFormattedString(msg, *parameters))
-	}
-
-	override fun fatalException(msg: String, exception: Exception, vararg parameters: Any) {
+	override fun exception(msg: String, exception: Exception, vararg parameters: Any) {
 		Log.wtf(tag, getFormattedString(msg, *parameters), exception)
-	}
-
-	override fun errorException(msg: String, exception: Exception, vararg parameters: Any) {
-		Log.e(tag, getFormattedString(msg, *parameters), exception)
 	}
 }
