+++ library/src/main/java/org/jellyfin/apiclient/interaction/connectionmanager/GetAvailableServersResponse.java
@@ -2,8 +2,8 @@ package org.jellyfin.apiclient.interaction.connectionmanager;
 
 import org.jellyfin.apiclient.interaction.ConnectionResult;
 import org.jellyfin.apiclient.interaction.Response;
+import org.jellyfin.apiclient.logging.ILogger;
 import org.jellyfin.apiclient.model.apiclient.ServerInfo;
-import org.jellyfin.apiclient.model.logging.ILogger;
 
 import java.util.ArrayList;
