+++ library/src/main/java/org/jellyfin/apiclient/interaction/discovery/ServerLocator.java
@@ -1,8 +1,8 @@
 package org.jellyfin.apiclient.interaction.discovery;
 
 import org.jellyfin.apiclient.interaction.Response;
+import org.jellyfin.apiclient.logging.ILogger;
 import org.jellyfin.apiclient.model.apiclient.ServerDiscoveryInfo;
-import org.jellyfin.apiclient.model.logging.ILogger;
 import org.jellyfin.apiclient.serialization.IJsonSerializer;
 
 import java.io.IOException;
@@ -46,7 +46,7 @@ public class ServerLocator implements IServerLocator {
                 c.send(sendPacket);
                 logger.debug("%s >>> Request packet sent to: 255.255.255.255 (DEFAULT)", getClass().getName());
             } catch (Exception e) {
-                logger.errorException("Error sending DatagramPacket", e);
+                logger.exception("Error sending DatagramPacket", e);
             }
 
             // Broadcast the message over all the network interfaces
@@ -69,7 +69,7 @@ public class ServerLocator implements IServerLocator {
                         DatagramPacket sendPacket = new DatagramPacket(sendData, sendData.length, broadcast, port);
                         c.send(sendPacket);
                     } catch (Exception e) {
-                        logger.errorException("Error sending DatagramPacket", e);
+                        logger.exception("Error sending DatagramPacket", e);
                     }
 
                     logger.debug("%s >>> Request packet sent to: %s; Interface: %s", getClass().getName(), broadcast.getHostAddress(), networkInterface.getDisplayName());
@@ -85,7 +85,7 @@ public class ServerLocator implements IServerLocator {
 
         } catch (Exception ex) {
 
-            logger.errorException("Error finding servers", ex);
+            logger.exception("Error finding servers", ex);
 
             response.onError(ex);
         }
