+++ library/src/main/java/org/jellyfin/apiclient/interaction/websocket/ApiWebSocket.java
@@ -3,9 +3,9 @@ package org.jellyfin.apiclient.interaction.websocket;
 import org.jellyfin.apiclient.interaction.ApiClient;
 import org.jellyfin.apiclient.interaction.ApiEventListener;
 import org.jellyfin.apiclient.interaction.EmptyResponse;
+import org.jellyfin.apiclient.logging.ILogger;
 import org.jellyfin.apiclient.model.apiclient.GeneralCommandEventArgs;
 import org.jellyfin.apiclient.model.entities.LibraryUpdateInfo;
-import org.jellyfin.apiclient.model.logging.ILogger;
 import org.jellyfin.apiclient.model.net.WebSocketMessage;
 import org.jellyfin.apiclient.model.session.BrowseRequest;
 import org.jellyfin.apiclient.model.session.GeneralCommandType;
