--- model/src/main/java/org/jellyfin/apiclient/model/logging/ILogManager.java
@@ -1,64 +1,0 @@
-package org.jellyfin.apiclient.model.logging;
-
-/** 
- Interface ILogManager
-*/
-public interface ILogManager
-{
-	/** 
-	 Gets or sets the log level.
-	 
-	 <value>The log level.</value>
-	*/
-	LogSeverity getLogSeverity();
-	void setLogSeverity(LogSeverity value);
-
-	/** 
-	 Gets or sets the exception message prefix.
-	 
-	 <value>The exception message prefix.</value>
-	*/
-	String getExceptionMessagePrefix();
-	void setExceptionMessagePrefix(String value);
-
-	/** 
-	 Gets the logger.
-	 
-	 @param name The name.
-	 @return ILogger.
-	*/
-	ILogger GetLogger(String name);
-
-	/** 
-	 Reloads the logger.
-	*/
-	void ReloadLogger(LogSeverity severity);
-
-	/** 
-	 Gets the log file path.
-	 
-	 <value>The log file path.</value>
-	*/
-	String getLogFilePath();
-
-	/** 
-	 Occurs when [logger loaded].
-	*/
-//C# TO JAVA CONVERTER TODO TASK: Events are not available in Java:
-//	event EventHandler LoggerLoaded;
-
-	/** 
-	 Flushes this instance.
-	*/
-	void Flush();
-
-	/** 
-	 Adds the console output.
-	*/
-	void AddConsoleOutput();
-
-	/** 
-	 Removes the console output.
-	*/
-	void RemoveConsoleOutput();
-}
